requirements.txt
assemblyai
python-dotenv
websockets
pyaudio
numpy
psutil
httpx
textblob
supabase
